+++ immuscreen/src/app/snp/[rsID]/eQTLs/page.tsx
@@ -34,6 +34,7 @@ export default function SNPeQTLs({
     }
   })
 
+
   return (
     <DataTable
       columns={[
